requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
datasets==4.1.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.35.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
instructor==1.11.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.27
lz4==4.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
openai==1.107.1
orjson==3.11.3
packaging==24.2
pandas==2.3.2
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.3.2
protobuf==5.29.5
protoc-gen-openapiv2==0.0.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.4
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
ragas==0.3.5
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scikit-network==0.33.3
scipy==1.16.1
sentence-transformers==5.1.0
sentry-sdk==2.37.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
